refactor: P0+P1 架构修复（core 反向依赖、contracts 扁平化、Any 类型、workers 反向依赖、God Object 拆分、dead code、测试清理）
P0 修复：
- dashboard_service 从 core 迁移到 workbench（消除 3 处顶层硬 import）
- bound_folder_scan_service 解耦（顶层 import → TYPE_CHECKING + 延迟）
- ImageRotation dead model 删除
- contracts 4 个过薄子包扁平化（repos/usecases/admin-workflows/assemblers）

P1 修复：
- 7 个工厂函数返回类型 Any → Protocol
- 4 个透传函数从 services/sms 移到 workers（消除反向依赖）
- folder_scan_service.py 拆分（1,138行 → 512+344+319）
- document_template_admin.py 拆分（1,029行 → 507+378+200）
- 18 个生成测试膨胀文件删除（~11,200 行）
