+++ backend/apps/contracts/services/contract/admin/__init__.py
@@ -1,13 +1,20 @@
 from __future__ import annotations
 
+from .case_creation_workflow import ContractCaseCreationWorkflow
+from .clone_workflow import ContractCloneWorkflow, plus_one_year_due_at
 from .contract_admin_document_service import ContractAdminDocumentService
 from .contract_admin_mutation_service import ContractAdminMutationService
 from .contract_admin_query_service import ContractAdminQueryService
 from .contract_admin_service import ContractAdminService
+from .filing_number_workflow import ContractFilingNumberWorkflow
 
 __all__ = [
     "ContractAdminDocumentService",
     "ContractAdminMutationService",
     "ContractAdminQueryService",
     "ContractAdminService",
+    "ContractCaseCreationWorkflow",
+    "ContractCloneWorkflow",
+    "ContractFilingNumberWorkflow",
+    "plus_one_year_due_at",
 ]
