+++ backend/apps/contracts/services/contract/contract_service_adapter.py
@@ -12,8 +12,8 @@ from apps.core.exceptions import NotFoundError
 from apps.core.interfaces import ContractDTO, ICaseService, LawyerDTO
 from apps.core.security.access_context import AccessContext
 
-from .assemblers.contract_details_assembler import ContractDetailsAssembler
-from .assemblers.contract_dto_assembler import ContractDtoAssembler
+from .query.contract_details_assembler import ContractDetailsAssembler
+from .query.contract_dto_assembler import ContractDtoAssembler
 from .contract_service import ContractService
 
 logger = logging.getLogger("apps.contracts")
