+++ backend/apps/contracts/services/contract/domain/access_policy.py
@@ -2,18 +2,28 @@
 
 from __future__ import annotations
 
+from collections.abc import Iterable
 from typing import TYPE_CHECKING, Any
 
 from django.db.models import Q, QuerySet
 
+from apps.contracts.models import Contract, ContractAssignment
 from apps.core.exceptions import PermissionDenied
 from apps.core.security import OrgAllowedLawyersMixin
 
-from ..repos.contract_access_repo import ContractAccessRepo
-
 if TYPE_CHECKING:
-    from apps.contracts.models import Contract
     from apps.core.security.access_context import AccessContext
+
+
+class ContractAccessRepo:
+    def has_assignment_access(self, *, contract_id: int, lawyer_ids: Iterable[int]) -> bool:  # pragma: no cover
+        lawyer_ids_list = list(lawyer_ids)
+        if not lawyer_ids_list:
+            return False
+        return ContractAssignment.objects.filter(contract_id=contract_id, lawyer_id__in=lawyer_ids_list).exists()
+
+    def has_case_assignment_access(self, *, contract_id: int, user_id: int) -> bool:  # pragma: no cover
+        return Contract.objects.filter(id=contract_id, cases__assignments__lawyer_id=user_id).exists()
 
 
 class ContractAccessPolicy(OrgAllowedLawyersMixin):
