+++ backend/apps/contracts/services/contract/query/__init__.py
@@ -1,5 +1,8 @@
 from __future__ import annotations
 
+from .contract_details_assembler import ContractDetailsAssembler
+from .contract_dto_assembler import ContractDtoAssembler
+from .contract_list_assembler import ContractListAssembler
 from .display_service import ContractDisplayService
 from .facade import ContractQueryFacade
 from .progress_service import ContractProgressService
@@ -7,7 +10,10 @@ from .service import ContractQueryService
 from .supplementary_agreement_query_service import SupplementaryAgreementQueryService
 
 __all__ = [
+    "ContractDetailsAssembler",
     "ContractDisplayService",
+    "ContractDtoAssembler",
+    "ContractListAssembler",
     "ContractProgressService",
     "ContractQueryFacade",
     "ContractQueryService",
