+++ backend/apps/contracts/services/contract/query/contract_details_assembler.py
