+++ backend/apps/contracts/services/contract/query/contract_dto_assembler.py
