+++ backend/apps/contracts/services/contract/query/contract_list_assembler.py
