+++ backend/apps/core/dependencies/business_contract.py
@@ -30,7 +30,7 @@ def build_contract_service() -> IContractService:
 
 def build_contract_service_with_deps(*, case_service: ICaseService, lawyer_service: ILawyerService) -> IContractService:
     from apps.contracts.services import ContractServiceAdapter
-    from apps.contracts.services.contract.usecases.composition import build_contract_service
+    from apps.contracts.services.contract.wiring import build_contract_service
 
     contract_service = build_contract_service(case_service=case_service, lawyer_service=lawyer_service)
     return ContractServiceAdapter(contract_service=contract_service)
