+++ backend/apps/workbench/services/__init__.py
@@ -1,11 +1,13 @@
 from .batch_service import BatchAnalysisService
 from .chat_service import WorkbenchChatService
+from .dashboard_service import DashboardService
 from .doc_extractor import DocTextExtractor
 from .message_service import WorkbenchMessageService
 from .session_service import WorkbenchSessionService
 
 __all__ = [
     "BatchAnalysisService",
+    "DashboardService",
     "DocTextExtractor",
     "WorkbenchChatService",
     "WorkbenchMessageService",
